VideoPlayerManager

+++ app/src/main/java/cn/api/message/ContentListByTypeQueryRequest.java
@@ -1,35 +1,41 @@
 package cn.api.message;
 
-public class ContentListByTypeQueryRequest 
-{
-	private int typeNum=0;
-	private String typeCode;
-	private int startId;
-	private int pageSize;
-	
-	public int getTypeNum() {
-		return typeNum;
-	}
-	public void setTypeNum(int typeNum) {
-		this.typeNum = typeNum;
-	}
-	public String getTypeCode() {
-		return typeCode;
-	}
-	public void setTypeCode(String typeCode) {
-		this.typeCode = typeCode;
-	}
-	public int getStartId() {
-		return startId;
-	}
-	public void setStartId(int startId) {
-		this.startId = startId;
-	}
-	public int getPageSize() {
-		return pageSize;
-	}
-	public void setPageSize(int pageSize) {
-		this.pageSize = pageSize;
-	}
-	
+public class ContentListByTypeQueryRequest {
+    private int typeNum = 0;
+    private String typeCode;
+    private long startId;
+    private int pageSize;
+
+    public int getTypeNum() {
+        return typeNum;
+    }
+
+    public void setTypeNum(int typeNum) {
+        this.typeNum = typeNum;
+    }
+
+    public String getTypeCode() {
+        return typeCode;
+    }
+
+    public void setTypeCode(String typeCode) {
+        this.typeCode = typeCode;
+    }
+
+    public long getStartId() {
+        return startId;
+    }
+
+    public void setStartId(long startId) {
+        this.startId = startId;
+    }
+
+    public int getPageSize() {
+        return pageSize;
+    }
+
+    public void setPageSize(int pageSize) {
+        this.pageSize = pageSize;
+    }
+
 }
